move kv:namespace out of index.tsx
