+++ packages/wrangler/src/index.tsx
@@ -14,20 +14,16 @@ import { DeprecationError } from "./errors";
 import { generateHandler, generateOptions } from "./generate";
 import { initHandler, initOptions } from "./init";
 import {
-	createKVNamespace,
 	deleteKVBulkKeyValue,
 	deleteKVKeyValue,
-	deleteKVNamespace,
 	getKVKeyValue,
 	getKVNamespaceId,
 	isKVKeyValue,
-	isValidKVNamespaceBinding,
 	listKVNamespaceKeys,
-	listKVNamespaces,
 	putKVBulkKeyValue,
 	putKVKeyValue,
 	unexpectedKVKeyValueProps,
-} from "./kv";
+} from "./kv/helpers";
 import { logger } from "./logger";
 import * as metrics from "./metrics";
 import { pages } from "./pages";
@@ -55,8 +51,9 @@ import {
 import { whoami } from "./whoami";
 
 import type { Config } from "./config";
-import type { KeyValue } from "./kv";
+import type { KeyValue } from "./kv/helpers";
 import type Yargs from "yargs";
+import { kvNamespace } from "./kv";
 
 export type ConfigPath = string | undefined;
 
@@ -152,7 +149,7 @@ export function getLegacyScriptName(
 
 // a helper to demand one of a set of options
 // via https://github.com/yargs/yargs/issues/1093#issuecomment-491299261
-function demandOneOfOption(...options: string[]) {
+export function demandOneOfOption(...options: string[]) {
 	return function (argv: Yargs.Arguments) {
 		const count = options.filter((option) => argv[option]).length;
 		const lastOption = options.pop();
@@ -452,160 +449,7 @@ function createCLIParser(argv: string[]) {
 		"kv:namespace",
 		"🗂️  Interact with your Workers KV Namespaces",
 		(namespaceYargs) => {
-			return namespaceYargs
-				.command(subHelp)
-				.command(
-					"create <namespace>",
-					"Create a new namespace",
-					(yargs) => {
-						return yargs
-							.positional("namespace", {
-								describe: "The name of the new namespace",
-								type: "string",
-								demandOption: true,
-							})
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							});
-					},
-					async (args) => {
-						await printWranglerBanner();
-
-						if (!isValidKVNamespaceBinding(args.namespace)) {
-							throw new CommandLineArgsError(
-								`The namespace binding name "${args.namespace}" is invalid. It can only have alphanumeric and _ characters, and cannot begin with a number.`
-							);
-						}
-
-						const config = readConfig(args.config as ConfigPath, args);
-						if (!config.name) {
-							logger.warn(
-								"No configured name present, using `worker` as a prefix for the title"
-							);
-						}
-
-						const name = config.name || "worker";
-						const environment = args.env ? `-${args.env}` : "";
-						const preview = args.preview ? "_preview" : "";
-						const title = `${name}${environment}-${args.namespace}${preview}`;
-
-						const accountId = await requireAuth(config);
-
-						// TODO: generate a binding name stripping non alphanumeric chars
-
-						logger.log(`🌀 Creating namespace with title "${title}"`);
-						const namespaceId = await createKVNamespace(accountId, title);
-						await metrics.sendMetricsEvent("create kv namespace", {
-							sendMetrics: config.send_metrics,
-						});
-
-						logger.log("✨ Success!");
-						const envString = args.env ? ` under [env.${args.env}]` : "";
-						const previewString = args.preview ? "preview_" : "";
-						logger.log(
-							`Add the following to your configuration file in your kv_namespaces array${envString}:`
-						);
-						logger.log(
-							`{ binding = "${args.namespace}", ${previewString}id = "${namespaceId}" }`
-						);
-
-						// TODO: automatically write this block to the wrangler.toml config file??
-					}
-				)
-				.command(
-					"list",
-					"Outputs a list of all KV namespaces associated with your account id.",
-					{},
-					async (args) => {
-						const config = readConfig(args.config as ConfigPath, args);
-
-						const accountId = await requireAuth(config);
-
-						// TODO: we should show bindings if they exist for given ids
-
-						logger.log(
-							JSON.stringify(await listKVNamespaces(accountId), null, "  ")
-						);
-						await metrics.sendMetricsEvent("list kv namespaces", {
-							sendMetrics: config.send_metrics,
-						});
-					}
-				)
-				.command(
-					"delete",
-					"Deletes a given namespace.",
-					(yargs) => {
-						return yargs
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to delete",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to delete",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							});
-					},
-					async (args) => {
-						await printWranglerBanner();
-						const config = readConfig(args.config as ConfigPath, args);
-
-						let id;
-						try {
-							id = getKVNamespaceId(args, config);
-						} catch (e) {
-							throw new CommandLineArgsError(
-								"Not able to delete namespace.\n" + ((e as Error).message ?? e)
-							);
-						}
-
-						const accountId = await requireAuth(config);
-
-						logger.log(`Deleting KV namespace ${id}.`);
-						await deleteKVNamespace(accountId, id);
-						logger.log(`Deleted KV namespace ${id}.`);
-						await metrics.sendMetricsEvent("delete kv namespace", {
-							sendMetrics: config.send_metrics,
-						});
-
-						// TODO: recommend they remove it from wrangler.toml
-
-						// test-mf wrangler kv:namespace delete --namespace-id 2a7d3d8b23fc4159b5afa489d6cfd388
-						// Are you sure you want to delete namespace 2a7d3d8b23fc4159b5afa489d6cfd388? [y/n]
-						// n
-						// 💁  Not deleting namespace 2a7d3d8b23fc4159b5afa489d6cfd388
-						// ➜  test-mf wrangler kv:namespace delete --namespace-id 2a7d3d8b23fc4159b5afa489d6cfd388
-						// Are you sure you want to delete namespace 2a7d3d8b23fc4159b5afa489d6cfd388? [y/n]
-						// y
-						// 🌀  Deleting namespace 2a7d3d8b23fc4159b5afa489d6cfd388
-						// ✨  Success
-						// ⚠️  Make sure to remove this "kv-namespace" entry from your configuration file!
-						// ➜  test-mf
-
-						// TODO: do it automatically
-
-						// TODO: delete the preview namespace as well?
-					}
-				);
+			return kvNamespace(namespaceYargs.command(subHelp));
 		}
 	);
